requirements.txt
asgiref==3.3.1
attrs==20.3.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
cached-property==1.5.2
certifi==2020.12.5
cffi
chardet==4.0.0
cryptography==2.9
distro==1.5.0
Django==3.1.4
django-ajax-selects==2.0.0
django-appconf==1.0.4
django-bootstrap-modal-forms==2.1.0
django-bootstrap-navbar==0.0.22
django-bootstrap4==2.3.1
django-braces==1.13.0
django-cors-headers==3.2.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.2
django-dynamic-formsets==0.0.8
django-enum-field==0.2
django-extensions==2.2.9
django-extra-views==0.13.0
django-filter==2.4.0
django-formset-js==0.5.0
django-integer-enum==0.1.0
django-jquery-js==3.1.1
django-model-utils==4.0.0
django-redis==4.12.1
django-rest-knox==4.1.0
django-select2==7.5.0
django-tables2==2.3.3
django-test-without-migrations==0.6
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
docker==5.0.0
docker-compose==1.29.1
dockerpty==0.4.1
docopt==0.6.2
drf-nested-field-multipart==0.0.3
drf-nested-multipart-parser==0.0.2
drf-rw-serializers==1.0.4
drf-writable-nested==0.6.2
idna==2.10
importlib-metadata==2.0.0
jsonschema==3.2.0
Markdown==3.3.3
paramiko==2.7.2
Pillow
psycopg2
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.17.0
python-enumeration==0.1
pytz==2019.3
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
simplejson==3.17.2
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.0
texttable==1.6.3
urllib3==1.26.4
websocket-client==0.58.0
zipp==3.4.0
